scripts/the-internet-herokuapp/requirements.txt
selenium
selenium-wire
behave
pyyaml
unittest
webdriver-manager
logging